refactor: navigation 리팩토링
