+++ src/components/navigation/consts/size.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_ITEM_SIZE = 40;
+export const DEFAULT_MAGNIFICATION = DEFAULT_ITEM_SIZE * 2;
+export const DEFAULT_DISTANCE = DEFAULT_MAGNIFICATION * 2;
