+++ src/components/navigation/index.ts
@@ -1,0 +1,1 @@
+export { Navigation } from './ui/navigation';
