+++ src/components/navigation/types/motion.ts
@@ -1,0 +1,7 @@
+import { MotionValue } from 'framer-motion';
+
+export type ItemMotionProps = {
+  magnification?: number;
+  distance?: number;
+  mousex?: MotionValue<number>;
+};
